+++ terraform/modules/roks_cluster/main.tf
@@ -18,6 +18,7 @@ terraform {
 
 module "cluster" {
   source = "./modules/cluster"
+  cluster_absent = var.cluster_absent
 
   ibmcloud_api_key = var.ibmcloud_api_key
   cluster_region   = var.ibmcloud_cluster_region